Memory manager getSize() improvements
- getSize() is renamed to getAllocatedSize() which returns allocated size of the block including header.
- Added getUsableSize() to return only usable size of memory block, without header.
- Added validateAndGetAllocatedSize() to first validate the pointer and then return allocated size.
- Added validateAndGetUsableSize() to first validate the pointer and then return usable size.
- All getSize() variants now returns long instead of int.
- getSize() variants delegates to validateAndGetSize() variants when assertions are enabled.
- Removed unused getPage() method.
- Improved and organized tests.

+++ hazelcast-enterprise/src/main/java/com/hazelcast/hidensity/HiDensityRecordAccessor.java
@@ -83,7 +83,7 @@ public interface HiDensityRecordAccessor<R extends HiDensityRecord>
      * @param memoryBlock {@link MemoryBlock} whose size will be calculated
      * @return the size of memory block
      */
-    int getSize(MemoryBlock memoryBlock);
+    long getSize(MemoryBlock memoryBlock);
 
     /**
      * Gets the size of given <code>address</code>.
